+++ R/util.R
@@ -1,7 +1,8 @@
+# ----------- Primary utilities ----------------------
 
 #' Utility functions
 #' 
-#' tabset, table layout, support for GIS shapefiles through library sf
+#' tabset, table layout, support for GIS shapefiles through library 'sf'
 #'  
 #' @param cmd utility command, see Details\cr
 #' @param js optional JavaScript function, default is NULL.\cr
@@ -275,13 +276,13 @@ ec.util <- function( ..., cmd='sf.series', js=NULL) {
       dfill <- ''
       if ('ppfill' %in% names(opts)) {
         dfill <- ifelse(is.null(opts$ppfill), 'null', paste0('"',opts$ppfill,'"'))
-        dfill <- paste0('ecfun.geofill=',dfill,';')
+        dfill <- paste0('ecf.geofill=',dfill,';')
       }
       out <- c( list(type= 'custom',
         coordinateSystem= 'leaflet',
         # set JS variables for riGeoJson() to work with
         renderItem= htmlwidgets::JS(paste("(params, api) => {",dfill,
-          " ecfun.geojson=",myGeojson,"; return riGeoJson(params, api); }")),
+          " ecf.geojson=",myGeojson,"; return riGeoJson(params, api); }")),
         data= if (is.null(opts$nid)) 
                 lapply(1:nrow(geojson$features), list)
               else 
@@ -478,7 +479,7 @@ body { padding: 10px; }
         icon= 'path://M5 5h5V3H3v7h2zm5 14H5v-5H3v7h7zm11-5h-2v5h-5v2h7zm-2-4h2V3h-7v2h5z',
         onclick= htmlwidgets::JS('function(){
                 tmp = this.ecModel.getOption();
-                ecfun.fscreen(tmp.echwid); 
+                ecf.fscreen(tmp.echwid); 
         }')
       ))
     },
@@ -549,10 +550,12 @@ body { padding: 10px; }
 #'  \item 'boxplot' = build dataset and source lists, see Details
 #' }
 #' @param header for dataset, to include the column names or not, default TRUE. Set it to FALSE for \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series.data}.\cr
-#' @param layout for boxplot, 'h' for horizontal(default) or 'v' for vertical layout\cr
-#' @param jitter for boxplot, value for \link[base]{jitter} of numerical values in second column, default 0 (no scatter). Adds scatter series on top of boxplot.\cr
-#' @param ... additional parameters for the boxplot _jitter_ (scatter) serie\cr
-#' Optional parameter _outliers_ for boxplot, a boolean to add outlier points (default FALSE)\cr
+#' @param ... optional parameters\cr
+#' Optional parameters for boxplot are:\cr
+#' * _layout_ 'h' for horizontal(default) or 'v' for vertical layout\cr
+#' * _outliers_ boolean to add outlier points (default FALSE)\cr
+#' * _jitter_ value for \link[base]{jitter} of numerical values in second column, default 0 (no scatter). Adds scatter series on top of boxplot.\cr
+#' 
 #' @return A list for _dataset.source_, _series.data_ or other lists:\cr
 #'   For boxplot - a named list, see Details and Examples \cr
 #'   For dendrogram & treePC - a tree structure, see format in \href{https://echarts.apache.org/en/option.html#series-tree.data}{tree data}
@@ -576,21 +579,21 @@ body { padding: 10px; }
 #' 
 #' hc <- hclust(dist(USArrests), "complete")
 #' ec.init(preset= FALSE,
-#'         series= list(list(
-#'           type= 'tree', orient= 'TB', roam= TRUE, initialTreeDepth= -1,
-#'           data= ec.data(hc, format='dendrogram'),
-#'           # layout= 'radial', symbolSize= ec.clmn(scale= 0.33),
-#'           ## exclude added labels like 'pXX', leaving only the originals
-#'           label= list(formatter= htmlwidgets::JS(
-#'             "function(n) { out= /p\\d+/.test(n.name) ? '' : n.name; return out;}"))
-#'         ))
+#'   series= list(list(
+#'     type= 'tree', orient= 'TB', roam= TRUE, initialTreeDepth= -1,
+#'     data= ec.data(hc, format='dendrogram'),
+#'     # layout= 'radial', symbolSize= ec.clmn(scale= 0.33),
+#'     ## exclude added labels like 'pXX', leaving only the originals
+#'     label= list(formatter= htmlwidgets::JS(
+#'       "function(n) { out= /p\\d+/.test(n.name) ? '' : n.name; return out;}"))
+#'   ))
 #' )
 #' 
 #' @importFrom utils tail
 #' @importFrom grDevices boxplot.stats
 #' @importFrom data.tree Aggregate
 #' @export
-ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ...) {
+ec.data <- function(df, format='dataset', header=FALSE, ...) {
   stopifnot('ec.data: expecting parameter df'= !missing(df))
   if (format=='dendrogram') { 
     stopifnot('ec.data: df should be hclust for dendrogram'= inherits(df, 'hclust'))
@@ -629,6 +632,8 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
     
   } 
   stopifnot('ec.data: df has to be data.frame'= inherits(df, 'data.frame'))
+  # add var for ec.clmn
+  .setColnm(colnames(df))
   
   if (format=='treePC') {
     # for sunburst,tree,treemap
@@ -694,6 +699,11 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
     datset <- lapply(tmp, \(x) list(value=unlist(unname(x))))
   } 
   else if (format=='boxplot') {
+    args <- list(...)
+    rady <- if ('ol.radius' %in% names(args)) args$ol.radius else NULL
+    jitter <- if ('jitter' %in% names(args)) args$jitter else 0
+    layout <- if ('layout' %in% names(args)) args$layout else 'h'
+    outliers <- if ('outliers' %in% names(args)) args$outliers else FALSE
     cn <- colnames(df)
     stopifnot('boxplot: df should have 2+ columns'= length(cn)>1)
     colas <- cn[1]
@@ -707,21 +717,20 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
     ttcol <- 1
     # category axis labels
     if (is.factor(df[[colas]]))
-        axe <- paste(levels(df[[colas]]), collapse="','")
+      axe <- paste(levels(df[[colas]]), collapse="','")
     else
-		    axe <- paste(sort(as.character(unique(df[[colas]]))), collapse="','")
+	    axe <- paste(sort(as.character(unique(df[[colas]]))), collapse="','")
 #     tbox <- "(p) => { d=p.data; out= '<b>'+p.seriesName+'</b>'+
 # '<br>min '+d[1].toFixed(3)+ '<br>Q1 '+d[2].toFixed(3)+
 # '<br>median '+d[3].toFixed(3)+ '<br>Q3 '+d[4].toFixed(3)+
 # '<br>max '+d[5].toFixed(3); return out; }"
     ttip <- c('Low', 'Q1', 'Q2', 'Q3', 'High')
-		args <- list(...)
-    
+
     if (!is.null(grpcol)) {   # grouped
       tmp <- df |> group_split()
       dataset <- lapply(tmp, \(dd) { 
         dv <- dd |> arrange(.data[[colas]]) |> 
-              dplyr::group_by(.data[[colas]]) |> group_split()
+              group_by(.data[[colas]]) |> group_split()
         src <- lapply(dv, \(vv) {
       	 if (nrow(vv)<5) stop(paste0('ec.data: insufficient data in group "'
       	                             ,grpcol,'"'), call.= FALSE)
@@ -740,20 +749,29 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
           type= 'boxplot', datasetIndex= i+length(tmp)-1) ))
       }
 	    axe <- paste0("function(v) { return ['",axe,"'][v]; }")
-  		if ('outliers' %in% names(args) && args$outliers) {
+	   
+  		if (outliers) {
   		  gcnt <- length(tmp)   # group count
         c2 <- gcnt+gcnt-1
         # new datasets with outliers
         dsotl <- lapply(gcnt:c2, \(x) list( fromDatasetIndex=x, fromTransformResult=1))
         serol <- lapply((gcnt*2):(gcnt+c2), \(x) 
-        	list(type='scatter', datasetIndex=x, name=series[[x-c2]]$name, z=4))
+      		list(type='custom', 
+      			  datasetIndex=x, renderItem= htmlwidgets::JS('riOutliers'),
+      			  encode= list(x=2, y=1),
+      			  name=series[[x-c2]]$name, z=4, 
+      			  itemStyle= list(borderDashOffset= rady)
+      	))
+			  if (layout=='v') {
+			    serol <- lapply(serol, \(ss) { 
+					  e <- ss$encode; ss$encode <- list(x=e$y, y=e$x); ss })
+			  }
         dataset <- append(dataset, dsotl)
-        series <- append(series, serol)
-  		}
+        series <- append(series, serol)  		}
     } 
     else {  # non-grouped
       bdf <- ungroup(df) |> arrange(.data[[colas]]) |>
-                 dplyr::group_by(across({colas})) |> group_split()
+                 group_by(across({colas})) |> group_split()
       src <- lapply(bdf, \(x) {x[[colb5]]})
       nms <- paste(unlist(lapply(bdf, \(x) {unique(x[[colas]])})), collapse="','")
       nms <- paste0("(p) => ['",nms,"'][p.value]")
@@ -766,17 +784,20 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
       series <- list(list(type='boxplot', name= 'boxplot', 
         datasetIndex= 1,
 			  encode= list(y= 1, x= c(2,3,4,5,6), tooltip=ttip)
-			))
+		  ))
       # default is horizontal, for vertical swap xAxis/yAxis category type
       if (layout=='v') {
         e <- series[[1]]$encode
         series[[1]]$encode <- list(x=e$y, y=e$x, tooltip=ttip) # swap x & y
       }
-		  axe <- paste0("function(v) { return ['",axe,"'][v-1]; }")
-  		if ('outliers' %in% names(args) && args$outliers) {
+		  axe <- "function(v) { return v;}"
+		  
+  		if (outliers) {
         # new dataset with outliers and serie for it
         dsotl <- list(list( fromDatasetIndex=1, fromTransformResult=1 ))
-        serol <- list(list(type='scatter', datasetIndex=2, name=series[[1]]$name, z=4))
+        serol <- list(list(type='scatter', 
+        		datasetIndex=2, name=series[[1]]$name, z=4, 
+        		itemStyle= list(borderDashOffset= rady) ))
         if (jitter==0) {
           serol[[1]] <-  .merlis(serol[[1]], args)
           series[[1]] <- .merlis(series[[1]], args)
@@ -795,7 +816,7 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
     
     if (jitter>0) {
   		tmp <- df |> arrange(.data[[colas]]) |> 
-  		        dplyr::group_by(.data[[colas]]) |> group_split()
+  		        group_by(.data[[colas]]) |> group_split()
   		mcyl <- lapply(tmp, \(x) unname(unlist(x[[colb5]])))
   		names(mcyl) <- sort(unique(df[[colas]]))
   		i <- 0.5
@@ -825,7 +846,7 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
     }
     return(list(dataset= dataset, series= series, xAxis=xaxis, yAxis=yaxis))
   } 
-  else {
+  else { # format=='names'
     datset <- tmp
   }  # format=='names'
   
@@ -838,23 +859,23 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
 #' Helper function to display/format data column(s) by index or name
 #' 
 #' @param col A single column index(number) or column name(quoted string), \cr
-#'    or a \link[base]{sprintf} format string for multiple indexes.\cr
+#'    or a \link[base]{sprintf} string template for multiple indexes.\cr
 #'    NULL(default) for charts with single values like tree, pie.\cr
-#'    'json' to display tooltip with all available values to choose from\cr 
+#'    'json' to display tooltip with all available values to choose from.\cr 
 #'    'log' to write all values in the JS console (F12) for debugging.\cr
-#'    if starts with _'function('_ or has _') =>'_ - to get result of JS function\cr
+#'    Can contain JS function starting with _'function('_ (or format _'(x) => {}'_).\cr
 #' @param ... Comma separated column indexes or names, only when \emph{col} is \emph{sprintf}. This allows formatting of multiple columns, as for a tooltip.\cr
 #' @param scale A positive number, multiplier for numeric columns. When scale is 0, all numeric values are rounded.
 #' @return A JavaScript code string (usually a function) marked as executable, see \link[htmlwidgets]{JS}.
 #'  
-#' @details This function is useful for attributes like formatter, color, symbolSize.\cr
-#' Column indexes are counted in R and start at 1.\cr
+#' @details This function is useful for attributes like formatter, color, symbolSize, label.\cr
+#' Column indexes are counted in R and start with 1.\cr
 #' Omit _col_ or use index -1 for single values in tree/pie charts, \emph{axisLabel.formatter} or \emph{valueFormatter}. See [ec.data] dendrogram example.\cr
-#' Use only column indexes when setting \emph{symbolSize}.\cr
 #' Column indexes are decimals for combo charts with multiple series, see [ecr.band] example. The whole number part is the serie index, the decimal part is the column index inside.\cr
 #' \emph{col} as sprintf has the same placeholder \emph{%@} for both column indexes or column names.\cr
 #' \emph{col} as sprintf can contain double quotes, but not single or backquotes.\cr
-#' Placeholders:\cr
+#' Template placeholders with formatting:\cr
+#' * \emph{%@} will display column value as-is.\cr
 #' * \emph{%L@} will display a number in locale format, like '12,345.09'.\cr
 #' * \emph{%LR@} rounded number in locale format, like '12,345'.\cr
 #' * \emph{%R@} rounded number, like '12345'.\cr
@@ -867,7 +888,7 @@ ec.data <- function(df, format='dataset', header=FALSE, jitter=0, layout='h', ..
 #' df |> dplyr::group_by(Species) |> ec.init(
 #'   series.param= list(label= list(show= TRUE, formatter= ec.clmn('emoji'))),
 #'   tooltip= list(formatter=
-#'     # ec.clmn with sprintf + multiple column indexes
+#'     # with sprintf template + multiple column indexes
 #'     ec.clmn('%M@ species <b>%@</b><br>s.len <b>%@</b><br>s.wid <b>%@</b>', 5,1,2))
 #' )
 #' 
@@ -880,8 +901,7 @@ ec.clmn <- function(col=NULL, ..., scale=1) {
     else scl <- paste0('return (parseFloat(c)*',scale,');') 
   }
   args <- list(...)
-  ret <- paste("let c=String(typeof x=='object' ? x.value : x);", scl)
-  remnl <- isTRUE(args$remnl)
+  ret <- paste("pos=[]; c= String(typeof x=='object' ? x.value : x);", scl)
   
   if (is.null(col)) {}   # for pie,sunburst
   else if (col=='log')
@@ -891,42 +911,63 @@ ec.clmn <- function(col=NULL, ..., scale=1) {
   else if (is.character(col) && (grepl(') =>', col) || 
                                  startsWith(col, 'function(')))
     return(htmlwidgets::JS(col))
-  else if (is.na(suppressWarnings(as.numeric(col)))) {   
-    if (length(args)==0) {  # col is solitary name
-      args <- c(col); col <- '%@'   # replace
-    }
-    # col is sprintf
-    spf <- "var sprintf= (template, values) => { let j=0;
-return template.replace(/%@|%L@|%LR@|%R@|%M@/g, (m) => {
-  if (m=='%@') return values[j++];
-  if (m=='%L@') return Number(values[j++]).toLocaleString();
-  if (m=='%LR@') return Math.round(Number(values[j++])).toLocaleString();
-  if (m=='%R@') return Math.round(Number(values[j++]));
+  else if (is.na(suppressWarnings(as.numeric(col)))) {
+  	
+	spf <- "var sprintf= (template, vals) => {
+j=0; if (template=='%@') return vals[j++];
+return template.replace(/%@|%L@|%LR@|%R@|%R2@|%M@/g, (m) => {
+  if (m=='%@') return vals[j++];
+  if (m=='%L@') return Number(vals[j++]).toLocaleString();
+  if (m=='%LR@') return Math.round(Number(vals[j++])).toLocaleString();
+  if (m=='%R@') return Math.round(Number(vals[j++]));
+  if (m=='%R2@') return Number(vals[j++]).toFixed(2);
   if (m=='%M@') return x.marker;
 }); };"
-    
-    tmp <- suppressWarnings(as.numeric(args) -1)
-    if (all(is.na(tmp))) {   
-      # multiple non-numeric strings = column names
-      t0 <- sapply(args, \(s) toString(paste0("x.data['", s,"']")) )
-      t0 <- paste(t0, collapse=',')
-      t1 <- paste(args, collapse='`,`')
-      # x.data = 1) object 2) array with x.dimensionNames 3) array only when dataset
-      ret <- paste0( spf, " if (!x.data) return `no data`; 
-let args=[`",t1,"`], vv=[",t0,"]; pos=[];
-if (x.dimensionNames && x.dimensionNames.length>0) 
-  pos= args.map(z => x.dimensionNames.indexOf(z));
-if (x.data.length)
-  vv= pos.map(p => x.data[p]);")
-    }   # col.names
-    else {   
-      #  multiple numeric, they could be in x, x.data, x.value OR x[].value[]
-      #  in combo-charts (ec.band), get decimal portion as .value index
-      tmp <- paste(tmp, collapse=',')
-      ret <- paste0( spf, "let ss=[",tmp,"];
-let vv= ss.map((e) => { 
+	if (length(args)==0) {  # col is solitary name
+		args <- col; col <- '%@'   # replace
+	}
+	# col is sprintf
+
+	tmp <- suppressWarnings(as.numeric(args) -1)
+	if (all(is.na(tmp))) {   
+		# multiple column names (non-numeric strings)
+		# to find position in colnames
+	  tmp <- .getColnm()
+		stopifnot("ec.clmn: colnames missing.
+    Use ec.clmn after ec.data and/or inside ec.init(df).
+    Otherwise use column indexes instead of names."= !is.null(tmp))
+		spf <- paste0(spf, " pos=['", paste(tmp, collapse="','"), "'];")
+		t0 <- sapply(args, \(s) toString(paste0("x.data['", s,"']")) )
+		t0 <- paste(t0, collapse=',')
+		t1 <- paste(args, collapse='`,`')
+		# ec.data(df,'values'): x is array for size, x.data.value for tooltip
+		# x.data = 
+		# 	1) object when ec.data('names')
+		# 	2) array only when ec.data('dataset') or df
+		#		3) x.data.value array with x.dimensionNames when ec.data('values')
+		# if series.dimensions=colnames(df) not set then x.dimensionNames may be wrong,
+		#   like ['lng', 'lat', 'value', 'value0',...]   console.log(x.dimensionNames);
+		# was if (x.dimensionNames && x.dimensionNames.length>0) pos= args.map(z => x.dimensionNames.indexOf(z)); else..
+		ret <- paste0( spf, " 
+aa= Array.isArray(x) ? x : x.data; tmp= null;
+if (aa && aa instanceof Object && !Array.isArray(aa)) {
+	tmp= Object.keys(aa); if (tmp.length==1 && tmp[0]=='value') aa= x.data.value;}
+if (tmp && tmp.length>1)
+	vv=[",t0,"];
+else {
+ if (!aa || !aa.length) return `no data`;
+ args= [`",t1,"`]; 
+ pos= args.map(z => pos.indexOf(z));
+ vv= pos.map(p => aa[p]); }")
+	}   # col.names
+	else {   
+		#  multiple numeric, they could be in x, x.data, x.value OR x[].value[]
+		#  in combo-charts (ec.band), get decimal portion as .value[] index
+		tmp <- paste(tmp, collapse=',')
+		ret <- paste0( spf, "ss=[",tmp,"];
+vv= ss.map((e) => { 
   if (e<0) return x.value ? x.value : x;
-  let i= Math.floor(e);
+  i= Math.floor(e);
   return x.value!=null ? x.value[i] : 
          x.data!=null  ? x.data[i] : 
          x[i]!=null    ? x[i] : `no data` });
@@ -936,27 +977,26 @@ if (vv.length > 0)
     f= Math.round(e % 1 *10) -1;
     return vv[idx].value[f];
   }); ")  # multi-series 1.2, 3.1
-    }  # col.indexes
+	}  # col.indexes
     
-    if (scale >0) ret <- paste(ret,"vv= vv.map(e => isNaN(e) | !e ? e : e*",scale,");")
-    if (scale==0) ret <- paste(ret,"vv= vv.map(e => isNaN(e) | !e ? e : Math.round(e));")
-    ret <- paste(ret, "let c = sprintf(`",col,"`, vv); return c; ")
+	if (scale >0) ret <- paste(ret,"vv= vv.map(e => isNaN(e) | !e ? e : e*",scale,");")
+	if (scale==0) ret <- paste(ret,"vv= vv.map(e => isNaN(e) | !e ? e : Math.round(e));")
+	# keep backwards-quotes for handling '\n'
+	ret <- paste0(ret, "c= sprintf(`",col,"`, vv); return c; ")
   } # col is string
   else {      # col is solitary numeric
     if (length(args) > 0)
       warning('col is numeric, others are ignored', call.=FALSE)
     col <- as.numeric(col) - 1   # from R to JS counting
     if (col >= 0)
-      ret <- paste0('let c = String(x.value!=null ? x.value[',col,
+      ret <- paste0('c= String(x.value!=null ? x.value[',col,
                     '] : x.data!=null ? x.data[',col,'] : x[',col,'] ); ',scl)
-  }
-  if (remnl)
-    ret <- gsub('\n', ' ', ret, fixed=T)
+  }  # col is solitary numeric
+  ret <- gsub('\t',' ', gsub('\n',' ', ret, fixed=T), fixed=T)
   htmlwidgets::JS(paste0('function(x) {', ret, '}'))
 }
 
 # ----------- Other utilities ----------------------
-#  more utilities in util.R
 
 
 #' Parallel Axis
@@ -985,9 +1025,9 @@ ec.paxis <- function(dfwt=NULL, cols=NULL, minmax=TRUE, ...) {
   if (inherits(dfwt, 'data.frame')) {
     coln <- colnames(dfwt)
     cfilter <- 1:length(coln)
-    if (dplyr::is.grouped_df(dfwt)) {
+    if (is.grouped_df(dfwt)) {
       # dont include grouping column (grnm)
-      grnm <- dplyr::group_vars(dfwt)[[1]]
+      grnm <- group_vars(dfwt)[[1]]
       cfilter <- cfilter[!cfilter==match(grnm, colnames(dfwt))]
     }
     idf <- dfwt