+++ tests/testthat/test-ec.clmn.R
@@ -1,59 +1,89 @@
-
+library(dplyr)
 test_that("ec.clmn with sprintf, column indexes and names", {
-  library(dplyr)
-  tmp <- data.frame(Species= as.vector(unique(iris$Species)), 
-                    color= c("#387e78","#eeb422","#d9534f"))
-  p <- iris |> inner_join(tmp) |> ec.init(
+  
+  # dataset + column indexes
+  p <- iris |> inner_join(data.frame(Species=unique(iris$Species), color= rainbow(3))) |>
+  ec.init(
     tooltip= list(formatter= ec.clmn('Petal Length %@, Width %@', 3,4)),
     series.param= list(itemStyle= list(color= ec.clmn(6)),
                        symbolSize= ec.clmn(3, scale=3))
   )
-  expect_true(p$x$opts$series[[1]]$itemStyle$color == "function(x) {let c = String(x.value!=null ? x.value[5] : x.data!=null ? x.data[5] : x[5] ); return c;}")
-  expect_s3_class( p$x$opts$series[[1]]$itemStyle$color, 'JS_EVAL')
-  expect_true(startsWith(p$x$opts$series[[1]]$symbolSize, "function(x) {let c = String(x.value!=null ?"))
-  expect_s3_class( p$x$opts$series[[1]]$symbolSize, 'JS_EVAL')
-  expect_match(p$x$opts$tooltip$formatter, "sprintf(` Petal Length %@, Width %@ `, vv)", fixed=TRUE)
-  expect_s3_class( p$x$opts$tooltip$formatter, 'JS_EVAL')
+  tmp <- p$x$opts$series[[1]]$itemStyle$color
+  expect_true(tmp == "function(x) {c= String(x.value!=null ? x.value[5] : x.data!=null ? x.data[5] : x[5] ); return c;}")
+  expect_s3_class(tmp, 'JS_EVAL')
+  tmp <-  p$x$opts$series[[1]]$symbolSize
+  expect_true(startsWith(tmp, "function(x) {c= String(x.value!=null ? x.value[2]"))
+  expect_s3_class(tmp, 'JS_EVAL')
+  tmp <- p$x$opts$tooltip$formatter
+  expect_match(tmp, "sprintf(`Petal Length %@, Width %@`, vv)", fixed=TRUE)
+  expect_s3_class(tmp, 'JS_EVAL')
   
+  # dataset + column names
+  p <- iris |> inner_join(data.frame(Species=unique(iris$Species), color= rainbow(3))) |>
+  ec.init(
+  	yAxis= list(axisLabel= list(formatter=ec.clmn('%R2@', -1, scale=0.5))),
+    series.param= list(
+      itemStyle= list(color= ec.clmn('color')),
+      symbolSize= ec.clmn('Petal.Length', scale=3)
+  	,tooltip= list(formatter= ec.clmn('Petal Length %@, Width %@', 'Petal.Length','Petal.Width'))
+    )
+    ,tooltip= list(s=T)  # enables series.tooltip
+    # does not pass thru ec.init(df) - cant use colnames!
+    # series= list(list( type='scatter',
+    #   itemStyle= list(color= ec.clmn('color')), # use 6 instead of 'color'
+    #   symbolSize= ec.clmn(3, scale=3)
+    # ))
+    #,tooltip= list(formatter= ec.clmn('Length %@, Width %@', 'Petal.Length','Petal.Width')) # use 3,4
+  )
+  tmp <- p$x$opts$series[[1]]
+  expect_match(tmp$tooltip$formatter, "vv=[x.data['Petal.Length'],x.data['Petal.Width']];", fixed=TRUE)
+  expect_match(tmp$itemStyle$color, "vv=[x.data['color']];", fixed=TRUE)
+  expect_match(tmp$itemStyle$color, "pos=['Sepal.Length'", fixed=TRUE)
+  expect_match(tmp$symbolSize, "vv=[x.data['Petal.Length']];", fixed=TRUE)
+  expect_match(p$x$opts$yAxis[[1]]$formatter, "ss=[-2]", fixed=TRUE)
   
+  # data + column names
   tmp <- data.frame(name=names(islands), value=islands) |> 
     filter(value>100) |> arrange(value)
-  p <- ec.init(
-    tooltip = list(formatter= ec.clmn("%@ <br> %L@", 'name','value')),
+  p <- ec.init( tooltip= list(s=TRUE),
     series = list(list(type= 'pie', data= ec.data(tmp, 'names'), 
-                  label= list(formatter= ec.clmn('value', scale=2))) )
+            label= list(formatter= ec.clmn('value', scale=2)),
+            tooltip= list(formatter= ec.clmn("%@ <br> %L@", 'name','value'))) )
   )
-  expect_equal(length(unlist(gregexpr('x.data', p$x$opts$tooltip$formatter ))), 5)
-  expect_true(grepl('[x.data.name,x.data.value]', p$x$opts$tooltip$formatter ))
-  expect_s3_class(p$x$opts$tooltip$formatter, 'JS_EVAL')
-  expect_equal(length(unlist(gregexpr('x.data', p$x$opts$series[[1]]$label$formatter ))), 4)
-  expect_is(p$x$opts$series[[1]]$label$formatter, 'JS_EVAL')
+  tmp <- p$x$opts$series[[1]]$tooltip$formatter
+  expect_equal(length(unlist(gregexpr('x.data', tmp ))), 4)
+  expect_match(tmp, "[x.data['name'],x.data['value']]", fixed=TRUE)
+  expect_s3_class(tmp, 'JS_EVAL')
+  tmp <- p$x$opts$series[[1]]$label$formatter
+  expect_equal(length(unlist(gregexpr('x.data', tmp ))), 3)
+  expect_is(tmp, 'JS_EVAL')
   
-  
+  # leaflet + mixed
   tmp <- quakes |> relocate('long') |>     # set order to lon,lat
     mutate(size= exp(mag)/20) |> head(100)  # add accented size
   p <- tmp |> ec.init(load='leaflet',
-    tooltip= list(formatter=ec.clmn('magnitude %@', 'mag')) # 'mag' or 4
+    tooltip= list(formatter= ec.clmn('magnitude %@', 4)) # 'mag' is 4th
   ) |> ec.upd({ 
-    series[[1]]$symbolSize = ec.clmn(6, scale=2)   # size column is 6th
+    series[[1]]$symbolSize <- ec.clmn('size', scale=2)
   })
-  expect_true(grepl("[x.data['mag']]", p$x$opts$tooltip$formatter, fixed=TRUE ))
-  expect_true(endsWith(p$x$opts$series[[1]]$symbolSize, '(parseFloat(c)*2);}'))
+  expect_true(grepl("ss=[3]", p$x$opts$tooltip$formatter, fixed=TRUE ))
+  expect_true(grepl("[x.data['size']]", p$x$opts$series[[1]]$symbolSize, fixed=TRUE ))
   
-  
-  tmp <- sort(islands)
-  tmp <- tmp[tmp>60]
-  tmp <- data.frame(name=names(tmp), value=as.character(unname(tmp)))
-  data <- ec.data(tmp, 'names')
-  p <- ec.init()
-  p$x$opts <- list(
+  # data + mixed
+  isl <- data.frame(name=names(islands), value=islands) |> arrange(desc(value)) |> 
+      head(13) |> mutate(colr=rainbow(13))
+  p <- ec.init(
     title = list(text = "Landmasses over 60,000 mi\u00B2", left = 'center'),
     tooltip = list(trigger='item', formatter=ec.clmn()),
-    series = list(type='pie', radius='50%', data= data, name='mi\u00B2')
+    series = list(list(type='pie', radius='50%', data= ec.data(isl, 'names'),
+                       itemStyle= list(color= ec.clmn('colr')),
+                       label=list(position='inside', 
+          formatter=ec.clmn("%@\n%@", 'name','value')) ))
   )
-  expect_equal("function(x) {let c=String(typeof x=='object' ? x.value : x); return c;}",
+  expect_equal("function(x) {pos=[]; c= String(typeof x=='object' ? x.value : x); return c;}",
                as.character(p$x$opts$tooltip$formatter) )
-  expect_equal(p$x$opts$series$data[[1]]$name, 'Celebes')
+  expect_equal(p$x$opts$series[[1]]$data[[1]]$name, 'Asia')
+  
   
   p <- ec.clmn('json', scale=NULL)
   expect_match(p, 'stringify', fixed=TRUE)